apply interface principle , dnt force class implenmnt a interface which that clss should not implement
