+++ src/AppBundle/InterfacePrinciple/AndroidWorker.php
@@ -1,0 +1,18 @@
+<?php
+namespace AppBundle\InterfacePrinciple;
+
+use AppBundle\InterfacePrinciple\WorkableInterface;
+use AppBundle\InterfacePrinciple\ManagableInterface;
+
+class AndroidWorker implements WorkableInterface, ManagableInterface
+{
+    public function work()
+    {
+        return  get_class($this) . ' '. __FUNCTION__ . ' ';
+    }
+
+    public function manage()
+    {
+        return $this->work();
+    }
+}
