+++ src/AppBundle/InterfacePrinciple/Captain.php
@@ -1,0 +1,12 @@
+<?php
+namespace AppBundle\InterfacePrinciple;
+
+class Captain
+{
+    public function manage(ManagableInterface $worker)
+    {
+        $log = '';
+        $log.=$worker->manage();
+        return $log;
+    }
+}
