+++ src/AppBundle/InterfacePrinciple/ManagableInterface.php
@@ -1,0 +1,7 @@
+<?php
+namespace AppBundle\InterfacePrinciple;
+
+interface ManagableInterface
+{
+    public function manage();
+}
