+++ src/AppBundle/InterfacePrinciple/SleepableInterface.php
@@ -1,0 +1,7 @@
+<?php
+namespace AppBundle\InterfacePrinciple;
+
+interface SleepableInterface
+{
+    public function sleep();
+}
