+++ src/AppBundle/Worker/HumanWorker.php
@@ -1,0 +1,8 @@
+<?php
+namespace AppBundle\InterfacePrinciple;
+
+use AppBundle\InterfacePrinciple\SleepableInterface;
+
+class HumanWorker implements Workable SleepableInterface{
+
+}
