+++ src/AppBundle/Worker/WorkableInterface.php
@@ -1,0 +1,7 @@
+<?php
+namespace AppBundle\InterfacePrinciple;
+
+interface WorkableInterface
+{
+    public function work();
+}
